requirements-dev.txt
-r requirements.txt
black==20.8b1
factory-boy==3.0.1
Faker==4.1.3
flake8==3.8.3
pytest==6.0.2
pytest-django==3.10.0
pytest-flake8==1.0.6
requests==2.24.0
